Worked on making key variables Global despite lack of support in C
Issues with lack of Global variable type continue to exist despite progress being made. Code should be fully compartmentalized by now and ready for testing of output into javascript and plotting. The only clear thing holding this back is the lack of Global variable support in C and issues with pre defining key variables at the start of the script. After this is overcome, the format of how the grid function is stored will need to be discerned before plotting.